feat: piilex v0.2.0 — 6 languages, 56 PII types, 5 frameworks, IDE plugins
N1:  Binary size optimization — feature flags (lang-web 5.1MB / lang-all 12MB)
N2:  CI pipeline validation — 178 tests, 5-platform builds, size checks
N3:  UX improvements — expanded --help, parse warnings with hints, progress bar
N4:  Security audit — serde_yml→serde_yaml_ng, input validation, SQL audit
N5:  VS Code Marketplace — VSIX packaging, extension commands, settings UI
N6:  npm/Homebrew distribution — dry-run verified, install.sh, formula
N7:  Landing page — SEO, OG tags, JSON-LD, Cloudflare Pages workflow
N8:  Community — issue/discussion templates, CONTRIBUTING.md, CODE_OF_CONDUCT.md
N9:  False positive reduction — phone regex fix, safe identifiers, context filters
N10: Dashboard integration — --upload flag, trend charts, PII distribution
N11: Custom rules YAML — custom sinks, allow/deny lists, path exceptions
N12: Performance — rayon parallel parsing (12K+ files/sec), memory tracking
N13: HIPAA/PCI-DSS — 7+6 rules, health/payment PII types
N14: Pre-commit hooks — --staged flag, piilex hook install/uninstall/status
N15: PostgreSQL migration — PG schema, PgBouncer guide, deployment docs
N16: Email notifications — findings alert, payment failure, weekly summary
N17: IntelliJ plugin — LSP client, settings, Gradle build, Marketplace workflow

+++ saas/api/src/models.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Team {
     pub id: String,
     pub name: String,
@@ -9,7 +9,7 @@ pub struct Team {
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct User {
     pub id: String,
     pub email: String,
@@ -19,40 +19,40 @@ pub struct User {
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Scan {
     pub id: String,
     pub team_id: String,
     pub user_id: Option<String>,
     pub project_name: String,
-    pub files_scanned: i64,
-    pub findings_count: i64,
-    pub critical_count: i64,
-    pub high_count: i64,
-    pub medium_count: i64,
-    pub low_count: i64,
+    pub files_scanned: i32,
+    pub findings_count: i32,
+    pub critical_count: i32,
+    pub high_count: i32,
+    pub medium_count: i32,
+    pub low_count: i32,
     pub frameworks: Option<String>,
-    pub duration_ms: i64,
+    pub duration_ms: i32,
     pub pii_type_summary: Option<String>,
     pub language_summary: Option<String>,
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct ScanFinding {
     pub id: String,
     pub scan_id: String,
     pub pii_type: String,
     pub severity: String,
     pub file_path: String,
-    pub line: i64,
+    pub line: i32,
     pub code_snippet: Option<String>,
     pub data_flow: Option<String>,
     pub framework_mappings: Option<String>,
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct ApiKey {
     pub id: String,
     pub user_id: String,
@@ -63,7 +63,7 @@ pub struct ApiKey {
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Subscription {
     pub id: String,
     pub team_id: String,
@@ -100,14 +100,14 @@ pub struct AuthResponse {
 #[derive(Debug, Deserialize)]
 pub struct CreateScanRequest {
     pub project_name: String,
-    pub files_scanned: i64,
-    pub findings_count: i64,
-    pub critical_count: i64,
-    pub high_count: i64,
-    pub medium_count: i64,
-    pub low_count: i64,
+    pub files_scanned: i32,
+    pub findings_count: i32,
+    pub critical_count: i32,
+    pub high_count: i32,
+    pub medium_count: i32,
+    pub low_count: i32,
     pub frameworks: Option<Vec<String>>,
-    pub duration_ms: i64,
+    pub duration_ms: i32,
     pub pii_type_summary: Option<serde_json::Value>,
     pub language_summary: Option<serde_json::Value>,
     pub findings: Option<Vec<CreateFindingRequest>>,
@@ -118,7 +118,7 @@ pub struct CreateFindingRequest {
     pub pii_type: String,
     pub severity: String,
     pub file_path: String,
-    pub line: i64,
+    pub line: i32,
     pub code_snippet: Option<String>,
     pub data_flow: Option<serde_json::Value>,
     pub framework_mappings: Option<serde_json::Value>,
